% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_expect_same_behavior.R
\name{expect_same_behavior}
\alias{expect_same_behavior}
\title{Expectation: do two calls behave the same ?}
\usage{
expect_same_behavior(object, expected, value = TRUE, output = TRUE,
  message = TRUE, warning = TRUE, failure = TRUE)
}
\arguments{
\item{object, expected}{Computation and value to compare it to.}

\item{value}{Wether to test if result is identical.}

\item{output}{Wether to test if printed output is identical.}

\item{message}{Wether to test if messages are identical.}

\item{warning}{Wether to test if warnings are identical.}

\item{failure}{Wether to test if both calls fail with the same error message.}

\item{desc}{Description, if not missing the call will forward `desc` to `test_that`.}
}
\value{
Returns `NULL` invisibly.
}
\description{
`expect_same_behavior()` tests if two calls trigger the same result,
 printed output, messages, warnings and failures. These can be tested separately
 or together.
}
\details{
`expect_same_behavior()` is wrapped by the `~+++++?` family of said dubious
 operators (powered by the package *doubt*)
}
